djangobackend/requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
Django==3.1.2
django-extensions==3.0.9
django-filter==2.4.0
djangorestframework==3.12.1
djangorestframework-jsonapi==4.0.0
gunicorn==20.0.4
inflection==0.5.1
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
wrapt==1.12.1
